The initial commit of Creatures 2 in Typescript
This project is intended to be a spiritual clone of Creatures 2, written in Typescript

The basic code is inspired directly by various source code leaks. It may be that things
don't work straight off or how they would do in the real game. There are many compromises
and compromises to come with how things are implemented.

I'm not sure how to handle game assets. I believe that many assets would be better served
if converted into modern formats, in terms of loading and programmatic management.

How this will effect the game or development of this project, it is yet to be seen.

The source code is provided without warranty of any kind, and some may be potentially
infringing, use at your own peril.

-  haxd 2022