Added code to handle background tile loading and infinite tiling
In order to match the visual appearance of the game, the background
should be able to be scrolled in either direction forever. This does
not include code to handle positioning of objects within the map just
yet. I want to integrate a bit more of the game logic before I continue.