Added React and Material UI for the game's user interface
The convention from Creatures 3 would be to implement the UI in
CAOS. I want to effectively make a fork of Creatures 2, even if
Creatures 3 is more modern. Maybe even because of that.

I've added some files as I started to map out how I want to parse CAOS
and also have begun to add some UI for a CAOS REPL. Should be neat and
also help me to build out the Macro system in general. I plan to use
the actions drawer as the main game UI. Maybe multiple drawers for each
kit? Or maybe modal windows that pop up (full screen if on mobile).

Nothing set in stone yet UI wise. I may end up making it as mobile
friendly as possible and abstracting away things that are useful for dev
and making things that are more useful for actually playing the game
easier to reach / interact with